dream_bot/requirements.txt
requests
python-telegram-bot
python-dotenv
dacite
boto3
mypy-boto3-dynamodb